Cache remapped mc jars + more refactoring
